Extract geotiff from zip file. Story #63
For works that have geo_mime_type of GeoTIFF, geo_works gem expects the
representative file to be the *.tif file, but we also want to attach the
zip file for easy user download.

For the attach_files method, I changed the file_names param to include
the file's path as well as the name instead of hard-coding the path
inside the method.  I made that change so I could use that same
attach_files method to attach temp files that were extracted from the
zip in a temporary directory (different directory from where the data
files are located).

This story only includes work to unzip files for GeoTIFF work.  If we
need to unzip files for other geo mime types, that will be future work.