This is a work-around to fix the problem described in story #50, where geotiff files weren't being properly published in GeoServer, so the raster image didn't appear in the geoblacklight front-end.
I overrode a method from rgeoserver gem to remove an element from the
XML document that was causing an error when POSTing to GeoServer.

I also added a rubocop exclusion for the new file because rubocop had a
lot of complaints.  Since I'm overriding a method from a gem, I wanted
the overridden code to be as identical as possible to the original code,
to make it easy for developers to compare the changes.